+++ src/statements/Stmt.ts
@@ -1,5 +1,6 @@
+import { Enviroment } from '../Enviroment'
 import { Literal } from '../Literal'
 
 export abstract class Stmt {
-  abstract execute(): Literal | null
+  abstract execute(enviroment: Enviroment): Literal | null
 }