+++ src/statements/index.ts
@@ -1,3 +1,4 @@
 export { Stmt } from './Stmt'
 export { StmtPrint } from './Print'
 export { StmtExpression } from './Expression'
+export { StmtVariable } from './Variable'